The two caller of cacheRelationship (addObject and valueUpdated) synchronized around the call with the relationship-lock.  This seems to be too restrictive.
Specifically in cacheRelationship, the call child.containerChange(old, parent), does its own relationship_lock locking, while it is walking all of its paths.

So I opened up the locks to not include child.containerChange(old, parent).

But there is a lurking bug, if two readers are following a path in reverse order. The current scheme for getting a path value will dead lock.